requirements.txt
loguru<=0.7.2
wandb>=0.12.9,<=0.13.4
matplotlib<=3.8.3
seaborn~=0.13.2
tabulate>=0.8.7,<=0.8.10
importlib-metadata==4.13.0
dask[distributed]
pydantic>=1.8.2,<=1.9.1
numpy>=1.21
pandas<=2.2.1
PyYAML<=6.0.1
pogema>=1.2.2